requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.20
botocore==1.17.20
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.5.0
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.42.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.1
django-cursor-pagination==0.1.4
django-debug-toolbar==2.2
django-elasticsearch-dsl==6.4.1
django-environ==0.4.5
django-excel==0.0.10
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-heroku==0.3.1
django-import-export==2.2.0
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
django-redis==4.11.0
django-request-logging==0.7.1
django-storages==1.9.1
django-webtest==1.9.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.15.2
elasticsearch==6.8.1
elasticsearch-dsl==6.4.0
et-xmlfile==1.0.1
fig==1.0.1
figaro==0.1
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
httpie==2.2.0
idna==2.9
isort==4.3.21
jdcal==1.4.1
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lml==0.0.9
Markdown==3.2.2
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.4
paramiko==2.7.1
phonenumbers==8.12.7
Pillow==7.2.0
promise==2.3
psycopg2==2.8.5
pycparser==2.20
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.6.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyperclip==1.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==3.13
pyzmail36==1.0.4
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
Rx==1.6.1
s3transfer==0.3.3
sendgrid==6.3.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sql==0.4.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.9
waitress==1.4.4
WebOb==1.8.6
websocket-client==0.57.0
WebTest==2.0.35
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
